add statistice to fill-pool.pl if it dies for some reason.
add Injection.conf, to inject data at a site and also create
soft-links or hard-links to populate a storage element with
pseudo-real files.

As configured, Injection.conf will inject 627,500 files in one
dataset, in blocks of 2500 files each. If the files are linked
to a source file of 2 GB, that corresponds to 1.3 PB of data.

Change the injection node and the dataset name if you want to run this
locally!